Fixed overlapping
Fixes #172 by adding DPIScalingCheck updating the label for `FavoriteWindowCombo` which was recently changed from `Label2` to `Label1`, as observed in the default steam skin cached settings page. For whatever reason none of the other labels were touched. This also integrates `TranslationLabel` and the, now, three dividers for that page instead of annihilating them.